+++ packages/express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/express
 
+## 0.1.120
+
+### Patch Changes
+
+- @vercel/node@5.9.2
+- @vercel/cervel@0.1.37
+
 ## 0.1.119
 
 ### Patch Changes